ADOP-1949: mesage update
