+++ src/main/java/uk/gov/hmcts/reform/adoption/adoptioncase/caseworker/event/CaseworkerSendOrReply.java
@@ -2,9 +2,6 @@ package uk.gov.hmcts.reform.adoption.adoptioncase.caseworker.event;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.apache.commons.collections4.CollectionUtils;
-import org.apache.commons.lang3.StringUtils;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.ccd.sdk.api.CCDConfig;
 import uk.gov.hmcts.ccd.sdk.api.CaseDetails;
@@ -13,7 +10,6 @@ import uk.gov.hmcts.ccd.sdk.api.callback.AboutToStartOrSubmitResponse;
 import uk.gov.hmcts.reform.adoption.adoptioncase.caseworker.event.page.SendOrReply;
 import uk.gov.hmcts.reform.adoption.adoptioncase.common.CaseEventCommonMethods;
 import uk.gov.hmcts.reform.adoption.adoptioncase.model.CaseData;
-import uk.gov.hmcts.reform.adoption.adoptioncase.model.MessageSendDetails;
 import uk.gov.hmcts.reform.adoption.adoptioncase.model.State;
 import uk.gov.hmcts.reform.adoption.adoptioncase.model.UserRole;
 import uk.gov.hmcts.reform.adoption.adoptioncase.model.access.Permissions;
@@ -23,20 +19,7 @@ import uk.gov.hmcts.reform.adoption.idam.IdamService;
 import uk.gov.hmcts.reform.idam.client.models.User;
 
 import javax.servlet.http.HttpServletRequest;
-import java.time.Instant;
-import java.time.LocalDateTime;
-import java.time.ZoneId;
-import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.UUID;
-import javax.servlet.http.HttpServletRequest;
 
-import static org.springframework.http.HttpHeaders.AUTHORIZATION;
-import static uk.gov.hmcts.reform.adoption.adoptioncase.search.CaseFieldsConstants.SEND_N_REPLY_USER_JUDGE;
-import static uk.gov.hmcts.reform.adoption.adoptioncase.search.CaseFieldsConstants.SEND_N_REPLY_DATE_FORMAT;
-import static uk.gov.hmcts.reform.adoption.adoptioncase.search.CaseFieldsConstants.COMMA;
-import static uk.gov.hmcts.reform.adoption.adoptioncase.search.CaseFieldsConstants.SEND_N_REPLY_USER_CASEWORKER;
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 
 @Slf4j
@@ -78,10 +61,7 @@ public class CaseworkerSendOrReply implements CCDConfig<CaseData, State, UserRol
     public AboutToStartOrSubmitResponse<CaseData, State> beforeStartEvent(CaseDetails<CaseData, State> details) {
         var caseData = details.getData();
         final User caseworkerUser = idamService.retrieveUser(request.getHeader(AUTHORIZATION));
-        CaseEventCommonMethods.prepareReplyMessageDynamicList(caseData);
-        caseData.setLoggedInUserRole(caseworkerUser.getUserDetails().getRoles()
-                                         .contains(UserRole.DISTRICT_JUDGE.getRole())
-                                         ? SEND_N_REPLY_USER_JUDGE : SEND_N_REPLY_USER_CASEWORKER);
+        CaseEventCommonMethods.prepareReplyMessageDynamicList(caseData,caseworkerUser);
         return AboutToStartOrSubmitResponse.<CaseData, State>builder()
             .data(caseData)
             .build();
@@ -93,19 +73,8 @@ public class CaseworkerSendOrReply implements CCDConfig<CaseData, State, UserRol
         final User caseworkerUser = idamService.retrieveUser(request.getHeader(AUTHORIZATION));
         var caseData = details.getData();
         CaseEventCommonMethods.updateMessageList(caseData,caseworkerUser);
-        caseData.setMessageAction(null);
-        caseData.setLoggedInUserRole(null);
         return AboutToStartOrSubmitResponse.<CaseData, State>builder()
             .data(caseData)
             .build();
-    }
-
-    private String getMessageReasonLabel(MessageSendDetails item) {
-        if (item.getMessageReasonList() != null && item.getMessageReasonList().getLabel() != null) {
-            return item.getMessageReasonList().getLabel();
-        } else if (item.getMessageReasonJudge() != null && item.getMessageReasonJudge().getLabel() != null) {
-            return item.getMessageReasonJudge().getLabel();
-        }
-        return null;
     }
 }
