+++ src/test/java/uk/gov/hmcts/reform/adoption/adoptioncase/common/CaseEventCommonMethodsTest.java
@@ -6,11 +6,12 @@ import uk.gov.hmcts.ccd.sdk.type.Document;
 import uk.gov.hmcts.ccd.sdk.type.DynamicListElement;
 import uk.gov.hmcts.ccd.sdk.type.ListValue;
 import uk.gov.hmcts.ccd.sdk.type.YesOrNo;
-import uk.gov.hmcts.reform.adoption.adoptioncase.model.MessageDocumentList;
 import uk.gov.hmcts.reform.adoption.adoptioncase.model.MessageSendDetails;
 import uk.gov.hmcts.reform.adoption.adoptioncase.model.CaseData;
-import uk.gov.hmcts.reform.adoption.adoptioncase.model.State;
+import uk.gov.hmcts.reform.adoption.adoptioncase.model.MessageDocumentList;
 import uk.gov.hmcts.reform.adoption.adoptioncase.model.SelectedMessage;
+import uk.gov.hmcts.reform.adoption.adoptioncase.model.State;
+import uk.gov.hmcts.reform.adoption.adoptioncase.model.UserRole;
 import uk.gov.hmcts.reform.adoption.document.model.AdoptionUploadDocument;
 import uk.gov.hmcts.reform.idam.client.models.User;
 import uk.gov.hmcts.reform.idam.client.models.UserDetails;
@@ -19,6 +20,7 @@ import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
+import java.util.Arrays;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static uk.gov.hmcts.reform.adoption.adoptioncase.common.CaseEventCommonMethods.prepareDocumentList;
@@ -67,7 +69,7 @@ public class CaseEventCommonMethodsTest {
         List<ListValue<MessageSendDetails>> listOfOpenMessage = new ArrayList<>();
         caseData.setListOfOpenMessages(caseData.archiveManageOrdersHelper(listOfOpenMessage,
                                                                           getListOfOpenMessages(UUID.randomUUID())));
-        prepareReplyMessageDynamicList(caseData);
+        prepareReplyMessageDynamicList(caseData, getCaseworkerUser());
         assertThat(caseData.getReplyMsgDynamicList()).isNotNull();
     }
 
@@ -79,7 +81,7 @@ public class CaseEventCommonMethodsTest {
         messageSendDetails.setMessageId("123e4567-e89b-12d3-a456-426614174000");
         messageSendDetails.setMessageSendDateNTime(LocalDateTime.now());
         messageSendDetails.setMessageStatus(MessageSendDetails.MessageStatus.OPEN);
-        messageSendDetails.setMessageReasonList(MessageSendDetails.MessageReason.LIST_A_HEARING);
+        messageSendDetails.setMessageReasonList(MessageSendDetails.MessageReason.ANNEX_A);
         caseData.setMessageSendDetails(messageSendDetails);
         CaseEventCommonMethods.updateMessageList(caseData, getCaseworkerUser());
         assertThat(caseData.getMessageAction()).isNull();
@@ -94,7 +96,7 @@ public class CaseEventCommonMethodsTest {
         messageSendDetails.setMessageId(uuid.toString());
         messageSendDetails.setMessageSendDateNTime(LocalDateTime.now());
         messageSendDetails.setMessageStatus(MessageSendDetails.MessageStatus.CLOSED);
-        messageSendDetails.setMessageReasonList(MessageSendDetails.MessageReason.LIST_A_HEARING);
+        messageSendDetails.setMessageReasonList(MessageSendDetails.MessageReason.ANNEX_A);
         caseData.setMessageSendDetails(messageSendDetails);
         List<ListValue<MessageSendDetails>> listOfOpenMessage = new ArrayList<>();
         caseData.setListOfOpenMessages(caseData.archiveManageOrdersHelper(listOfOpenMessage,
@@ -102,7 +104,7 @@ public class CaseEventCommonMethodsTest {
         var selectedMessage = new SelectedMessage();
         selectedMessage.setReplyMessage(YesOrNo.YES);
         caseData.setSelectedMessage(selectedMessage);
-        prepareReplyMessageDynamicList(caseData);
+        prepareReplyMessageDynamicList(caseData, getCaseworkerUser());
         caseData.getReplyMsgDynamicList().setValue(new DynamicListElement(uuid, "Test"));
         CaseEventCommonMethods.updateMessageList(caseData, getCaseworkerUser());
         assertThat(caseData.getMessageAction()).isNull();
@@ -113,7 +115,7 @@ public class CaseEventCommonMethodsTest {
         messageSendDetails.setMessageId(uuid.toString());
         messageSendDetails.setMessageSendDateNTime(LocalDateTime.now());
         messageSendDetails.setMessageStatus(MessageSendDetails.MessageStatus.OPEN);
-        messageSendDetails.setMessageReasonList(MessageSendDetails.MessageReason.LIST_A_HEARING);
+        messageSendDetails.setMessageReasonList(MessageSendDetails.MessageReason.ANNEX_A);
         return  messageSendDetails;
     }
 
@@ -181,6 +183,7 @@ public class CaseEventCommonMethodsTest {
     private User getCaseworkerUser() {
         UserDetails userDetails = UserDetails
             .builder()
+            .roles(Arrays.asList(UserRole.DISTRICT_JUDGE.getRole()))
             .forename("testFname")
             .surname("testSname")
             .build();
