+++ src/test/java/uk/gov/hmcts/reform/adoption/adoptioncase/common/CommonPageBuilderTest.java
@@ -9,16 +9,21 @@ import uk.gov.hmcts.ccd.sdk.type.ListValue;
 import uk.gov.hmcts.reform.adoption.adoptioncase.model.CaseData;
 import uk.gov.hmcts.reform.adoption.adoptioncase.model.MessageSendDetails;
 import uk.gov.hmcts.reform.adoption.adoptioncase.model.State;
+import uk.gov.hmcts.reform.adoption.adoptioncase.model.UserRole;
 import uk.gov.hmcts.reform.adoption.document.model.AdoptionUploadDocument;
+import uk.gov.hmcts.reform.idam.client.models.User;
+import uk.gov.hmcts.reform.idam.client.models.UserDetails;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
+import java.util.Arrays;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static uk.gov.hmcts.reform.adoption.adoptioncase.common.CaseEventCommonMethods.prepareReplyMessageDynamicList;
 import static uk.gov.hmcts.reform.adoption.adoptioncase.common.CommonPageBuilder.sendMessageMidEvent;
+import static uk.gov.hmcts.reform.adoption.testutil.TestConstants.TEST_AUTHORIZATION_TOKEN;
 import static uk.gov.hmcts.reform.adoption.testutil.TestDataHelper.caseData;
 
 public class CommonPageBuilderTest {
@@ -54,7 +59,7 @@ public class CommonPageBuilderTest {
         var uuid = UUID.randomUUID();
         caseData.getData().setListOfOpenMessages(caseData.getData().archiveManageOrdersHelper(listOfOpenMessage,
                                                                           getListOfOpenMessages(uuid)));
-        prepareReplyMessageDynamicList(caseData.getData());
+        prepareReplyMessageDynamicList(caseData.getData(), getCaseworkerUser());
         caseData.getData().getReplyMsgDynamicList().setValue(new DynamicListElement(uuid, "test"));
         AboutToStartOrSubmitResponse<CaseData, State> response = sendMessageMidEvent(caseData, caseData);
         assertThat(response.getData().getSelectedMessage()).isNotNull();
@@ -73,7 +78,7 @@ public class CommonPageBuilderTest {
         messageSendDetails.setMessageId(uuid.toString());
         messageSendDetails.setMessageSendDateNTime(LocalDateTime.now());
         messageSendDetails.setMessageStatus(MessageSendDetails.MessageStatus.OPEN);
-        messageSendDetails.setMessageReasonList(MessageSendDetails.MessageReason.LIST_A_HEARING);
+        messageSendDetails.setMessageReasonList(MessageSendDetails.MessageReason.ANNEX_A);
         messageSendDetails.setMessageUrgencyList(MessageSendDetails.MessageUrgency.HIGH);
         messageSendDetails.setMessageText("test");
         return  messageSendDetails;
@@ -122,5 +127,16 @@ public class CommonPageBuilderTest {
         uploadDocument.setName("testdoc5.jpg");
         uploadDocument.setDocumentLink(new Document());
         return  uploadDocument;
+    }
+
+    private User getCaseworkerUser() {
+        UserDetails userDetails = UserDetails
+            .builder()
+            .forename("testFname")
+            .roles(Arrays.asList(UserRole.DISTRICT_JUDGE.getRole()))
+            .surname("testSname")
+            .build();
+
+        return new User(TEST_AUTHORIZATION_TOKEN, userDetails);
     }
 }
